add stuff

+++ src/main/kotlin/com/iulia/gardener/controller/FeaturesController.kt
@@ -1,6 +1,7 @@
 package com.iulia.gardener.controller
 
-import com.iulia.gardener.integration.impl.FeaturesConfigurationDtoService
+import com.iulia.gardener.service.impl.FeaturesConfigurationDtoService
+import com.iulia.gardener.service.impl.UserService
 import org.openapitools.gardener.api.FeaturesApi
 import org.openapitools.gardener.model.FeaturesConfigurationDto
 import org.springframework.http.HttpStatus
@@ -14,11 +15,23 @@ import javax.validation.Valid
 import javax.validation.constraints.NotNull
 
 @Controller
-class FeaturesController( var service: FeaturesConfigurationDtoService): FeaturesApi {
+class FeaturesController(var service: FeaturesConfigurationDtoService,
+                         var userService: UserService) : FeaturesApi {
 
     override fun deleteFeatureConfiguration(@PathVariable(value = "id") id: UUID, @NotNull @RequestParam(required = true, value = "userToken") userToken: String): ResponseEntity<Unit> {
-        service.delete(id)
-        return ResponseEntity(HttpStatus.OK)
+        var resource = service.getOne(id)
+        if (resource?.userId == null && userService.isAdmin(userToken)) {
+            service.delete(id)
+            return ResponseEntity(HttpStatus.OK)
+        } else if (resource?.userId != null) {
+            var requester = userService.getUserByToken(userToken).id
+            var owner = resource.userId
+            if (requester == owner) {
+                service.delete(id)
+                return ResponseEntity(HttpStatus.OK)
+            }
+        }
+        return ResponseEntity(HttpStatus.UNAUTHORIZED)
     }
 
     override fun getFeatureConfiguration(id: UUID): ResponseEntity<FeaturesConfigurationDto> {
